requirements.txt
python-telegram-bot==21.6
runwayml==3.12.0
python-dotenv==1.0.1
aiohttp
